Try to fix issue https://github.com/astronomer/astronomer-cosmos/actions/runs/21001735120/job/60373303976
